requirements.txt.txt
fastapi
uvicorn
pymongo
python-dotenv
